+++ presentation/src/main/java/com/cheocharm/presentation/ui/MainActivity.kt
@@ -1,25 +1,52 @@
 package com.cheocharm.presentation.ui
 
+import android.Manifest
 import android.content.Context
 import android.graphics.Rect
+import android.os.Build
 import android.os.Bundle
 import android.view.MotionEvent
 import android.view.View
 import android.view.inputmethod.InputMethodManager
 import android.widget.EditText
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.annotation.RequiresApi
 import androidx.core.view.isVisible
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.setupWithNavController
 import com.cheocharm.presentation.R
 import com.cheocharm.presentation.base.BaseActivity
 import com.cheocharm.presentation.databinding.ActivityMainBinding
+import com.google.android.gms.location.FusedLocationProviderClient
+import com.google.android.gms.location.LocationServices
 import com.google.android.gms.maps.SupportMapFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
     private var mapFragment: SupportMapFragment? = null
+    private lateinit var fusedLocationClient: FusedLocationProviderClient
 
+    @RequiresApi(Build.VERSION_CODES.N)
+    private val locationPermissionRequest = registerForActivityResult(
+        ActivityResultContracts.RequestMultiplePermissions()
+    ) { permissions ->
+        when {
+            permissions.getOrDefault(Manifest.permission.ACCESS_FINE_LOCATION, false) -> {
+                fusedLocationClient =
+                    LocationServices.getFusedLocationProviderClient(this)
+            }
+            permissions.getOrDefault(Manifest.permission.ACCESS_COARSE_LOCATION, false) -> {
+                fusedLocationClient =
+                    LocationServices.getFusedLocationProviderClient(this)
+            }
+            else -> {
+                // TODO: 위치 권한을 얻지 못했을 때
+            }
+        }
+    }
+
+    @RequiresApi(Build.VERSION_CODES.N)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -39,6 +66,13 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
                 binding.bottomNavMain.visibility = View.VISIBLE
             }
         }
+
+        locationPermissionRequest.launch(
+            arrayOf(
+                Manifest.permission.ACCESS_FINE_LOCATION,
+                Manifest.permission.ACCESS_COARSE_LOCATION
+            )
+        )
 
         mapFragment =
             supportFragmentManager.findFragmentById(R.id.fragment_main_map) as? SupportMapFragment
@@ -71,4 +105,6 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
     }
 
     fun getMap(): SupportMapFragment? = mapFragment
+
+    fun getLocationClient(): FusedLocationProviderClient = fusedLocationClient
 }
