+++ presentation/src/main/java/com/cheocharm/presentation/ui/home/HomeFragment.kt
@@ -4,7 +4,6 @@ import android.Manifest
 import android.content.pm.PackageManager
 import android.os.Bundle
 import android.view.View
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.viewModels
 import com.cheocharm.presentation.R
@@ -16,8 +15,6 @@ import com.cheocharm.presentation.common.SOUTH_KOREA_ZOOM_LEVEL
 import com.cheocharm.presentation.common.toLatLng
 import com.cheocharm.presentation.databinding.FragmentHomeBinding
 import com.cheocharm.presentation.ui.MainActivity
-import com.google.android.gms.location.FusedLocationProviderClient
-import com.google.android.gms.location.LocationServices
 import com.google.android.gms.maps.CameraUpdateFactory
 import com.google.android.gms.maps.model.LatLng
 import dagger.hilt.android.AndroidEntryPoint
@@ -26,39 +23,15 @@ import dagger.hilt.android.AndroidEntryPoint
 class HomeFragment : BaseFragment<FragmentHomeBinding>(R.layout.fragment_home) {
     private val homeViewModel: HomeViewModel by viewModels()
 
-    private lateinit var fusedLocationClient: FusedLocationProviderClient
-
-    private val locationPermissionRequest = registerForActivityResult(
-        ActivityResultContracts.RequestMultiplePermissions()
-    ) { permissions ->
-        when {
-            permissions.getOrDefault(Manifest.permission.ACCESS_FINE_LOCATION, false) -> {
-                fusedLocationClient =
-                    LocationServices.getFusedLocationProviderClient(requireActivity())
-            }
-            permissions.getOrDefault(Manifest.permission.ACCESS_COARSE_LOCATION, false) -> {
-                fusedLocationClient =
-                    LocationServices.getFusedLocationProviderClient(requireActivity())
-            }
-            else -> {
-                // TODO: 위치 권한을 얻지 못했을 때
-            }
-        }
-    }
+    private lateinit var mainActivity: MainActivity
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
         binding.viewmodel = homeViewModel
 
-        (activity as MainActivity).setMapVisible(true)
-
-        locationPermissionRequest.launch(
-            arrayOf(
-                Manifest.permission.ACCESS_FINE_LOCATION,
-                Manifest.permission.ACCESS_COARSE_LOCATION
-            )
-        )
+        mainActivity = activity as MainActivity
+        mainActivity.setMapVisible(true)
 
         val mapFragment = (activity as MainActivity).getMap()
         mapFragment?.getMapAsync { map ->
@@ -68,7 +41,7 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(R.layout.fragment_home) {
                         Manifest.permission.ACCESS_FINE_LOCATION
                     ) == PackageManager.PERMISSION_GRANTED
                 ) {
-                    fusedLocationClient.lastLocation.addOnSuccessListener { location ->
+                    mainActivity.getLocationClient().lastLocation.addOnSuccessListener { location ->
                         map.moveCamera(
                             CameraUpdateFactory.newLatLngZoom(
                                 location.toLatLng(),
